clippy_dev: Split `rename` and `deprecate` out of `update_lints`
